+++ idealingua/idealingua-runtime-rpc-http4s/src/main/scala/com/github/pshirshov/izumi/idealingua/runtime/rpc/http4s/WithHttp4sClient.scala
@@ -12,7 +12,7 @@ trait WithHttp4sClient {
 
   //  protected def clientMarshallers: IRTClientMarshallers
   //
-  class ClientDispatcher(baseUri: Uri, codec: IRTCodec)
+  class ClientDispatcher(baseUri: Uri, codec: IRTClientMultiplexor)
     extends Dispatcher with IRTZioResult {
 
     private val client: CIO[Client[CIO]] = Http1Client[CIO]()
