+++ idealingua/idealingua-runtime-rpc-scala/src/main/scala/com/github/pshirshov/izumi/idealingua/runtime/rpc/IRTClientMultiplexor.scala
@@ -1,10 +1,9 @@
 package com.github.pshirshov.izumi.idealingua.runtime.rpc
 
-import io.circe.ParsingFailure
-import io.circe.parser._
+import io.circe.parser.parse
 import scalaz.zio.IO
 
-class IRTCodec(clients: Set[IRTWrappedClient]) extends IRTZioResult {
+class IRTClientMultiplexor(clients: Set[IRTWrappedClient]) extends IRTZioResult {
   val codecs: Map[IRTMethodId, IRTMarshaller] = clients.flatMap(_.allCodecs).toMap
 
   def encode(input: IRTMuxRequest[Product]): IO[Throwable, String] = {
@@ -32,32 +31,6 @@ class IRTCodec(clients: Set[IRTWrappedClient]) extends IRTZioResult {
 
       case None =>
         IO.terminate(new IRTMissingHandlerException(s"No codec for $method, input=$input", input, None))
-    }
-  }
-}
-
-class IRTMultiplexor[C](list: Set[IRTWrappedService[C]]) extends IRTZioResult {
-  val services: Map[IRTServiceId, IRTWrappedService[C]] = list.map(s => s.serviceId -> s).toMap
-
-
-  def doInvoke(body: String, context: C, toInvoke: IRTMethodId): Either[ParsingFailure, Option[IO[Throwable, String]]] = {
-    for {
-      parsed <- _root_.io.circe.parser.parse(body)
-    } yield {
-      for {
-        service <- services.get(toInvoke.service)
-        method <- service.allMethods.get(toInvoke)
-      } yield {
-        for {
-          decoded <- method.marshaller.decodeRequest(IRTJsonBody(toInvoke, parsed))
-          casted <- IO.syncThrowable(decoded.value.asInstanceOf[method.signature.Input])
-          result <- IO.syncThrowable(method.invoke(context, casted))
-          safeResult <- result
-          encoded <- IO.syncThrowable(method.marshaller.encodeResponse(IRTResBody(safeResult)).noSpaces)
-        } yield {
-          encoded
-        }
-      }
     }
   }
 }
