+++ idealingua/idealingua-runtime-rpc-scala/src/main/scala/com/github/pshirshov/izumi/idealingua/runtime/rpc/IRTMethodWrapper.scala
@@ -1,8 +1,5 @@
 package com.github.pshirshov.izumi.idealingua.runtime.rpc
 
-import io.circe.{DecodingFailure, Json}
-import scalaz.zio.IO
-
